+++ FF1Lib/ItemLocations.cs
@@ -265,18 +265,18 @@ namespace FF1Lib
         public static TreasureChest Unused15 = new TreasureChest(0x31FF, nameof(Unused15), 0, 0, isUnused: true);
 
         public static MapObject KingConeria = new MapObject(ObjectId.King, MapLocation.ConeriaCastle, Item.Bridge);
-        public static MapObject Princess = new MapObject(ObjectId.Princess2, MapLocation.ConeriaCastle, Item.Lute);
+        public static MapObject Princess = new MapObject(ObjectId.Princess2, MapLocation.ConeriaCastle, Item.Lute, requiredSecondLocation:MapLocation.TempleOfFiends);
         public static MapObject Matoya = new MapObject(ObjectId.Matoya, MapLocation.MatoyasCave, Item.Herb, AccessRequirement.Crystal, requiredItemTrade: Item.Crystal);
         public static MapObject Bikke = new MapObject(ObjectId.Bikke, MapLocation.Pravoka, Item.Ship, useVanillaRoutineAddress:true);
         // Assumption is made that if you have access to the Elf Prince you also have access to the Elf Doc
         public static MapObject ElfPrince = new MapObject(ObjectId.ElfPrince, MapLocation.ElflandCastle, Item.Key, AccessRequirement.Herb, ObjectId.ElfDoc);
         public static MapObject Astos = new MapObject(ObjectId.Astos, MapLocation.NorthwestCastle, Item.Crystal, AccessRequirement.Crown, useVanillaRoutineAddress: true);
-        public static MapObject Sarda = new MapObject(ObjectId.Sarda, MapLocation.SardasCave, Item.Rod, requiredGameEventFlag:ObjectId.Vampire);
+        public static MapObject Sarda = new MapObject(ObjectId.Sarda, MapLocation.SardasCave, Item.Rod, requiredGameEventFlag:ObjectId.Vampire, requiredSecondLocation:MapLocation.EarthCave);
         public static MapObject CanoeSage = new MapObject(ObjectId.CanoeSage, MapLocation.CresentLake, Item.Canoe, requiredItemTrade:Item.EarthOrb);
         public static MapObject CubeBot =  new MapObject(ObjectId.CubeBot, MapLocation.Waterfall, Item.Cube);
         public static MapObject Fairy =  new MapObject(ObjectId.Fairy, MapLocation.Gaia, Item.Oxyale, AccessRequirement.Bottle);
         // Assumption is made that if you have the slab and access to Lefein then you also have access to Unne
-        public static MapObject Lefein = new MapObject(ObjectId.Lefein, MapLocation.Lefein, Item.Chime, AccessRequirement.Slab, ObjectId.Unne);
+        public static MapObject Lefein = new MapObject(ObjectId.Lefein, MapLocation.Lefein, Item.Chime, AccessRequirement.Slab, ObjectId.Unne, requiredSecondLocation: MapLocation.Melmond);
         public static MapObject Nerrick = new MapObject(ObjectId.Nerrick, MapLocation.DwarfCave, Item.Canal, AccessRequirement.Tnt, requiredItemTrade: Item.Tnt); 
         public static MapObject Smith = new MapObject(ObjectId.Smith, MapLocation.DwarfCave, Item.Xcalber, AccessRequirement.Adamant, requiredItemTrade: Item.Adamant);
